Add UART to the software. The minibot can now send 'hello world' over BT. Add the two programming manuals to the documents. Also fix the system clock

+++ software/peripherals/sys/sys.h
@@ -12,8 +12,8 @@ typedef void (*isr_handler_t)();
 void sys_setup_external_clock(void);
 
 // Delay the given amount of time
-void delay_us(int us);
-void delay_ms(int ms);
+void delay_us(unsigned long us);
+void delay_ms(unsigned long ms);
 
 
 #endif
